# scripts/deployment/openapi.yaml
openapi: 3.1.0
info:
  title: The New Fuse API
  version: 1.0.0
  description: |
    Comprehensive API documentation for The New Fuse platform - a multi-agent orchestration system with blockchain integration, workflow automation, and AI-powered capabilities.

    ## Features
    - **Agent Management**: Create, manage, and orchestrate AI agents
    - **Workflow Engine**: Build and execute complex workflows
    - **Chat System**: Real-time messaging and communication
    - **Web3 Integration**: Wallet management, transactions, and smart accounts (ERC-4337)
    - **MCP Protocol**: Model Context Protocol for agent communication
    - **Authentication**: JWT-based secure authentication
    - **Rate Limiting**: Tiered rate limiting based on user roles

    ## Authentication
    Most endpoints require JWT authentication. Include the Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Base URLs
    - Development: `http://localhost:3001/api`
    - API Gateway: `http://localhost:4000/api/v1`
    - Production: `https://api.thenewfuse.com/api`

  contact:
    name: The New Fuse API Support
    url: https://thenewfuse.com/support
    email: support@thenewfuse.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Development API Server
  - url: http://localhost:4000/api/v1
    description: API Gateway (Development)
...
    description: Production API Server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token obtained from /auth/login
  schemas:
...
    AgentType:
      type: string
      enum:
        - BASIC
        - CHAT
        - WORKFLOW
        - TASK
        - ASSISTANT
        - ANALYSIS
        - CONVERSATIONAL
        - IDE_EXTENSION
        - API
    AgentStatus:
      type: string
      enum:
        - active
        - inactive
        - paused
        - busy
        - error
    AgentCapability:
      type: string
      enum:
        - natural_language
        - code_generation
        - data_analysis
        - image_processing
        - web_scraping
        - api_integration
        - workflow_execution
    CreateAgentDto:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          example: "Customer Support Bot"
        type:
          $ref: "#/components/schemas/AgentType"
        description:
          type: string
          maxLength: 500
          example: "Handles customer inquiries and support tickets"
        systemPrompt:
          type: string
          example: "You are a helpful customer support assistant"
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/AgentCapability"
        configuration:
          type: object
          additionalProperties: true
          example:
            maxConcurrent: 10
            responseTimeout: 30
    UpdateAgentDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        systemPrompt:
          type: string
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/AgentCapability"
        configuration:
          type: object
          additionalProperties: true
        status:
          $ref: "#/components/schemas/AgentStatus"
    AgentResponseDto:
      type: object
      properties:
        id:
          type: string
          example: "agent123"
        name:
          type: string
        type:
          $ref: "#/components/schemas/AgentType"
        status:
          $ref: "#/components/schemas/AgentStatus"
        description:
          type: string
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/AgentCapability"
        configuration:
          type: object
          additionalProperties: true
        metrics:
          type: object
          properties:
            uptime:
              type: number
              format: float
            avgResponseTime:
              type: number
              format: float
            satisfaction:
              type: number
              format: float
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastActiveAt:
          type: string
          format: date-time

    # Chat Schemas
...
paths:
...
  /agents:
    get:
      tags:
        - Agents
      summary: Get all agents
      operationId: getAgents
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/AgentType"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/AgentStatus"
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentResponseDto"

    post:
      tags:
        - Agents
      summary: Create a new agent
      operationId: createAgent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentDto"
      responses:
        "201":
          description: Agent created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponseDto"
        "400":
          description: Invalid agent data
...
  /agents/{id}:
    get:
      tags:
        - Agents
      summary: Get agent by ID
      operationId: getAgentById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponseDto"
        "404":
          description: Agent not found

    put:
      tags:
        - Agents
      summary: Update agent
      operationId: updateAgent
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentDto"
      responses:
        "200":
          description: Agent updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponseDto"
        "404":
          description: Agent not found

    delete:
      tags:
        - Agents
      summary: Delete agent
      operationId: deleteAgent
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Agent deleted successfully
        "404":
          description: Agent not found
...